+++ src/withContext.js
@@ -45,18 +45,4 @@ function withContext(ContextAPIs, mapContextToProps) {
   }
 }
 
-/**
- * Temporary workaround to fix react-redux issue with forwardRef.
- * This workaround will be removed on v3.
- *
- * Issue: https://github.com/pgarciacamou/react-context-consumer-hoc/issues/6
- */
-function noRef(ContextAPIs, mapContextToProps) {
-  return ComposedComponent => {
-    const Component = withContext(ContextAPIs, mapContextToProps)(ComposedComponent)
-    return props => (<Component {...props} />)
-  }
-}
-
-withContext.noRef = noRef
 export default withContext